+++ CHANGELOG.md
@@ -175,4 +175,5 @@ Added some world leak fixes for:
 Added patches for Forestry:
 - Allow Analyzer with Errors
 - Allow Partially Filled Containers
+
 Add IC2 Explosion Protection